APP/Streamlit/requirements.txt
numpy
streamlit
pymysql
scipy
pandas
python-dotenv